Make Ot.Var.Dim a class instead of a literal type.
