+++ doc/pages/references/ot/var.mdx
@@ -12,11 +12,15 @@ export default Layout
 
 An immutable, opaque union type of <R s="number"/> and an internal class implements a quantity that varies under OpenType Variation rules. Use <R s={Ot.Var.Ops}/> to create and manipulate them.
 
-### Type <Decl s={Ot.Var.Dim} />
+### Class <Decl s={Ot.Var.Dim} />
 
 An object type representing a “variation dimension” in a vaiable font. Its object identity is significant, and is used to distinguish dimensions of a variable font's deisgn space.
 
 Class <R s={Ot.Fvar.Axis}/> in <R s={Ot.Fvar}/> contains such dimension object, with extra properties for naming, identification, etc.
+
+#### Constructor
+
+ * <Method ctor s={Ot.Var.Dim.Constructor} args={{tag:Tag,minValue:number,defaultValue:number,maxValue:number}}/>
 
 #### Properties
