+++ packages/test-util/src/test-variance.ts
@@ -2,24 +2,9 @@ import { OtVar } from "@ot-builder/variance";
 
 export namespace TestVariance {
     export function createWellKnownAxes() {
-        const wght: OtVar.Dim = {
-            tag: "wght",
-            min: 100,
-            default: 400,
-            max: 900
-        };
-        const wdth: OtVar.Dim = {
-            tag: "wdth",
-            min: 25,
-            default: 100,
-            max: 200
-        };
-        const opsz: OtVar.Dim = {
-            tag: "opsz",
-            min: 6,
-            default: 12,
-            max: 72
-        };
+        const wght = new OtVar.Dim("wght", 100, 400, 900);
+        const wdth = new OtVar.Dim("wdth", 25, 100, 200);
+        const opsz = new OtVar.Dim("opsz", 6, 12, 72);
 
         return { wght, wdth, opsz };
     }
