+++ packages/variance/src/interface/dimension.ts
@@ -1,12 +1,18 @@
 import { Tag } from "@ot-builder/primitive";
 
-export interface VarianceDim {
+export class VarianceDim {
     readonly tag: Tag;
     readonly min: number;
     readonly default: number;
     readonly max: number;
-}
 
-export interface VarianceDimMapper<T extends VarianceDim> {
-    (from: VarianceDim): T;
+    // This property is added to prevent literals being recognized as Dim
+    protected readonly __dimensionType = 1;
+
+    constructor(tag: Tag, min: number, defaultVal: number, max: number) {
+        this.tag = tag;
+        this.min = min;
+        this.default = defaultVal;
+        this.max = max;
+    }
 }
